blog_system-master/requirements.txt
beautifulsoup4==4.6.0
Django==2.1.1
html5lib==0.999999999
Pillow==5.0.0
PyMySQL==0.9.2
pytz==2017.2
Werkzeug==0.14.1
